requirements.txt
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-search-nspkg==3.0.1
azure-cognitiveservices-search-websearch==1.0.0
azure-common==1.1.22
azure-nspkg==3.0.2
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
msrest==0.6.8
oauthlib==3.0.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
Werkzeug==0.15.5
